consumer/requirements.txt
kafka-python==2.0.2
pymysql==1.1.0
python-dotenv==1.0.0
